修改优化

+++ controller/article.js
@@ -1,5 +1,7 @@
 const formidable = require('formidable')
 const articleModel = require('../mongodb/models/article.js')
+const tagModel = require('../mongodb/models/tag.js')
+const dbFn = require('../mongodb/method.js')
 const util = require('../util')
 
 module.exports = {
@@ -24,12 +26,17 @@ module.exports = {
         const clicks = util.randNum(45,260)
         const stars = util.randNum(2,24)
         const article_index = await util.getId('article_id')
-        const date_string = new Date().toLocaleString()
+        const create_date = util.formatDateTime()
+        const tags = fields.tags || []
+        dbFn.addOrUpdateTag(tags)
+        if (!fields.cover) {
+          delete fields.cover
+        }
         await articleModel.create({
           ...fields,
-          date_string,
           clicks,
           stars,
+          create_date,
           id: util.PrefixInteger(article_index, 6)
         })
         res.send({
@@ -46,26 +53,60 @@ module.exports = {
     })
   },
   async delete(req, res) {
-    if(req.query.id){
-      try{
-        await articleModel.deleteOne({ id: req.query.id});
-        res.send({
-          status:1,
-          message: '文章删除成功'
-        })
-      }catch(err){
-        res.send({
-          status:0,
-          message: '删除出现错误'
-        })
-        throw Error(err)
-      }  
+    if (req.query.id) {
+      if (req.query.remove) {
+        // 真删除
+        try {
+          await articleModel.deleteOne({ id: req.query.id })
+          res.send({
+            status:1,
+            message: '文章已从回收站删除！'
+          })
+        } catch (error) {
+          res.send({
+            status:0,
+            message: '从回收站删除出现错误'
+          })
+          throw Error(err)
+        }
+      } else {
+        // 加入回收站
+        try{
+          const article = await articleModel.findOneAndUpdate({ id: req.query.id }, { deleted: true });
+          const tags = article.tags
+          tags.forEach( async item => {
+            await tagModel.updateOne({ name: item }, {$inc: {useNum: -1}})
+          })
+          res.send({
+            status:1,
+            message: '文章删除成功'
+          })
+        }catch(err){
+          res.send({
+            status:0,
+            message: '删除出现错误'
+          })
+          throw Error(err)
+        }
+      }
     }else{
       res.send({
         status:0,
         message: '未获取到文章id，请重新操作'
       })
     }
+  },
+  async recovery(req, res) {
+    const id = req.query.id
+    const article = await articleModel.findOne({ id })
+    const tags = article.tags || []
+    article.deleted = false
+    article.save()
+    dbFn.addOrUpdateTag(tags)
+    res.send({
+      status: 1,
+      message: '恢复成功'
+    })
   },
   async update (req, res) {
     const form = new formidable.IncomingForm()
@@ -79,7 +120,18 @@ module.exports = {
         })
       }
       try {
-        await articleModel.update({id:id},content)
+        const article = await articleModel.findOne({ id })
+        const oldTagArr = article.tags
+        const newTagarr = content.tags
+        await article.update(content)
+        const newAdd = util.filter2Array(newTagarr, oldTagArr)
+        const subtract = util.filter2Array(oldTagArr, newTagarr)
+        console.log('n:', newAdd)
+        console.log('s:', subtract)
+        dbFn.addOrUpdateTag(newAdd)
+        subtract.forEach(async x => {
+          await tagModel.updateOne({name : x}, {$inc: {useNum: -1}})
+        })
         res.send({
           status: 1,
           message: '文章更新成功'
@@ -99,7 +151,7 @@ module.exports = {
     const skip = limit*(page-1)
     if(req.query.deleted){
       try {
-        const articleList = await articleModel.find({"deleted":true}, '-_id -__v -html -create_date').sort({'create_date':-1}).limit(limit).skip(skip)
+        const articleList = await articleModel.find({"deleted":true}, '-_id -__v -html').sort({'create_date':-1}).limit(limit).skip(skip)
         const articleTotal = await articleModel.countDocuments({"deleted":true})
         res.send({
           status: 1,
@@ -112,7 +164,7 @@ module.exports = {
       }
     }else{
       try {
-        const articleList = await articleModel.find({"deleted":{$ne: true}}, '-_id -__v -html -create_date').sort({'create_date':-1}).limit(limit).skip(skip)
+        const articleList = await articleModel.find({"deleted":{$ne: true}}, '-_id -__v -html').sort({'create_date':-1}).limit(limit).skip(skip)
         const articleTotal = await articleModel.countDocuments({"deleted":{$ne: true}})
         res.send({
           status: 1,
@@ -127,7 +179,7 @@ module.exports = {
   },
   async hot(req,res) {
     try {
-      const articleList = await articleModel.find({}, '-_id -__v -html -create_date').sort({'clicks':-1}).limit(5)
+      const articleList = await articleModel.find({"deleted":{$ne: true}}, '-_id -__v -html').sort({'clicks':-1}).limit(5)
       res.send({
         status: 1,
         data: articleList,
@@ -139,8 +191,7 @@ module.exports = {
   },
   async handleClick(req, res) {
     const id = req.params.article_id
-    // const data = await articleModel.findOne({id}, '-__v -_id -create_date')
-    const data = await articleModel.findOneAndUpdate({id},{$inc:{ clicks:1 }},{projection: { "__v" : 0, "_id" : 0, "create_date": 0 }})
+    const data = await articleModel.findOneAndUpdate({id},{$inc:{ clicks:1 }},{projection: { "__v" : 0, "_id" : 0 }})
     res.send({
       status: 1,
       data,
@@ -151,13 +202,13 @@ module.exports = {
     const { id, type }= req.query
     try {
       if(type === 'star') {
-        await articleModel.update({id}, {$inc:{ stars: 1}})
+        await articleModel.updateOne({id}, {$inc:{ stars: 1}})
         res.send({
           status: 1,
           message: '加心成功'
         })
       }else{
-        await articleModel.update({id}, {$inc:{ stars: -1}})
+        await articleModel.updateOne({id}, {$inc:{ stars: -1}})
         res.send({
           status: 1,
           message: '取消加心成功'
